refactor: unify rate limiter
